+++ src/genomicSimulation-init.c
@@ -19,7 +19,7 @@ R_CallMethodDef calledMethods[] = {
 	{"SXP_see_optimal_haplotype", (DL_FUNC) &SXP_see_optimal_haplotype, 2},
 	{"SXP_get_optimal_GEBV", (DL_FUNC) &SXP_get_optimal_GEBV, 2},
 	{"SXP_get_minimal_GEBV", (DL_FUNC) &SXP_get_minimal_GEBV, 2},
-	{"SXP_see_group_data", (DL_FUNC) &SXP_see_group_data, 4},
+	{"SXP_see_group_data", (DL_FUNC) &SXP_see_group_data, 5},
 	{"SXP_see_group_gene_data", (DL_FUNC) &SXP_see_group_gene_data, 3},
 	{"SXP_see_existing_groups", (DL_FUNC) &SXP_see_existing_groups, 1},
 	{"SXP_save_GEBVs", (DL_FUNC) &SXP_save_GEBVs, 4},